refactor: Fix UByte bitwise operation type conversion in CsvBitmap and JsonBitmap

+++ src/commonMain/kotlin/borg/trikeshed/parse/csv/CsvBitmap.kt
@@ -85,7 +85,7 @@ object CsvBitmap : DocumentBitmap {
                     val b: UByte = if ((maskedSoFar % 2).z)
                         (input[inputY][inputX / 2].toUInt() shr 4).toUByte()
                     else
-                        input[inputY][inputX / 2] and 0b0000_1111U
+                        (input[inputY][inputX / 2] and 0b0000_1111U).toUByte()
 
                     val maskBits = b.toUInt() shr 2 and 0x3u
